Update Stream - Kth Largest.cpp

+++ Data Structures/Heaps/Stream - Kth Largest.cpp
@@ -16,7 +16,7 @@ class KthLargest
     void add(ll val)
     {
         pq.push(val);
-        if (pq.size() > k)
+        if ((ll) pq.size() > k)
         {
             pq.pop();
         }
